+++ ContentPatcher/docs/release-notes.md
@@ -10,6 +10,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 
 -->
 ## Upcoming release
+* Added `AddNpcWarps` field in [`EditMap` patches](https://github.com/Pathoschild/StardewMods/blob/develop/ContentPatcher/author-guide.md#editmap).
 * Fixed some cases where a local token on an `Include` patch wasn't available in the included patches.
 
 ## 2.5.3